refactored SwiftPackageFileAnalyzer
